Add Lua Scripting in Autonomous
Users can write scripts in lua to control their robot while in the AUTON
state.
A menu appears in the SETUP state that allows the user to select a
script (or no script).
Refactored LuaTest into LuaScriptLoader.
Added small delay to the lua interpreter after every statement so that
it doesn't overload the actual game.
Refactored several instanceof of "ship" to "robot".
Miscellaneous fixes.